an_authorised_user_may_not_delete_goals_which_they_do_not_have_permission_to_delete

+++ app/Http/Controllers/Api/Goal/GoalController.php
@@ -66,7 +66,11 @@ class GoalController extends Controller
      */
     public function destroy(Goal $goal)
     {
-        $goal->delete();
-        return response()->json(['error' => false], 204);
+        if (auth()->id() === $goal->user_id) {
+            $goal->delete();
+            return response()->json(['error' => false], 204);
+        }
+
+        return response()->json(['error' => true], 403);
     }
 }
